+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Types of changes
 
 - Support for accessing the I/Os of Kunbus digital and analog I/O modules for the Revolution Pi
 - Device driver classes for CETONI I/O channels
+- Increase required Python version to 3.8 because in 3.7 the implementation of `ThreadPoolExecutor` in the standard library does not reuse idle threads leading to an ever increasing number of threads which eventually causes blocking of the server(s) on Raspberry Pis
 
 ### Changed